+++ samples/react-comparer/src/services/OneDriveServices/IGetListDataAsStreamRequest.ts
@@ -1,0 +1,7 @@
+export interface IGetListDataAsStreamRequest {
+  parameters:{
+    RenderOptions :number;
+    ViewXml:string;
+    AllowMultipleValueFilterForTaxonomyFields:boolean;
+  };
+}